backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
transformers==4.11.3
torch==1.9.1
flask-cors==3.0.10
pymupdf
python-docx